+++ src/components/index.js
@@ -20,10 +20,14 @@ import RpcTester from './Network/RPC/RpcTester'
 import FormSendTx from './Tx/SendTx'
 import TxHistory from './Tx/TxHistory'
 
+import AccountList from './Wallet/AccountList'
 import AccountItem from './Wallet/AccountItem'
 import TokenListForItem from './Wallet/TokenListForItem'
+import NavbarItem from './Wallet/NavbarItem'
+import NetworkStatus from './Wallet/NetworkStatus'
 
 export {
+  AccountList,
   AccountItem,
   Button,
   Checkbox,
@@ -39,7 +43,9 @@ export {
   FormSendTx,
   Identicon,
   InputPassword,
+  NavbarItem,
   NetworkChooser,
+  NetworkStatus,
   NodeInfo,
   NodeInfoBox,
   TxHistory,